requirements.txt
google-adk
mcp
python-dotenv
streamlit